Use float for version number such as CentOS 6.3.

+++ source-builder/sb/linux.py
@@ -69,14 +69,14 @@ def load():
 
     # Works for LSB distros
     distro = platform.dist()[0]
-    distro_ver = int(platform.dist()[1])
+    distro_ver = float(platform.dist()[1])
 
     # Non LSB - fail over to issue
     if distro == '':
         try:
             issue = open('/etc/issue').read()
             distro = issue.split(' ')[0]
-            distro_ver = int(issue.split(' ')[2])
+            distro_ver = float(issue.split(' ')[2])
         except:
             pass
